+++ Source/LibationWinForms/AvaloniaUI/Views/ProductsDisplay2.axaml
@@ -11,7 +11,7 @@
 		<DataGrid Name="productsGrid" AutoGenerateColumns="False" Items="{Binding GridEntries}">
 			<DataGrid.Columns>
 				
-				<controls:DataGridCheckBoxColumnExt IsVisible="False" Header="Remove" IsThreeState="True" IsReadOnly="False" CanUserSort="True" Binding="{Binding Remove, Mode=TwoWay}" Width="70" SortMemberPath="Remove"/>
+				<controls:DataGridCheckBoxColumnExt IsVisible="True" Header="Remove" IsThreeState="True" IsReadOnly="False" CanUserSort="True" Binding="{Binding Remove, Mode=TwoWay}" Width="70" SortMemberPath="Remove"/>
 
 				<DataGridTemplateColumn CanUserSort="True" Width="75" Header="Liberate" SortMemberPath="Liberate">
 					<DataGridTemplateColumn.CellTemplate>
